chore: split up infractions into multiple files.
