+++ interactions/infractions/infractions.go
@@ -1,20 +1,16 @@
 package infractions
 
 import (
-	"context"
 	"fmt"
 	"log/slog"
 	"math"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/disgoorg/disgo/discord"
 	"github.com/disgoorg/disgo/handler"
-	"github.com/disgoorg/json"
 	"github.com/disgoorg/snowflake/v2"
 
-	"github.com/NLLCommunity/heimdallr/interactions"
 	"github.com/NLLCommunity/heimdallr/model"
 	"github.com/NLLCommunity/heimdallr/utils"
 )
@@ -40,470 +36,6 @@ func Register(r *handler.Mux) []discord.ApplicationCommandCreate {
 
 // pageSize is the size of one page of infractions
 const pageSize = 5
-
-//░█░█░█▀█░█▀▄░█▀█░░░█▀▀░█▀█░█▄█░█▄█░█▀█░█▀█░█▀▄
-//░█▄█░█▀█░█▀▄░█░█░░░█░░░█░█░█░█░█░█░█▀█░█░█░█░█
-//░▀░▀░▀░▀░▀░▀░▀░▀░░░▀▀▀░▀▀▀░▀░▀░▀░▀░▀░▀░▀░▀░▀▀░
-
-// WarnCommand lets moderators issue warnings to users.
-var WarnCommand = discord.SlashCommandCreate{
-	Name: "warn",
-	NameLocalizations: map[discord.Locale]string{
-		discord.LocaleNorwegian: "advar",
-	},
-	Description: "Warn a user.",
-	DescriptionLocalizations: map[discord.Locale]string{
-		discord.LocaleNorwegian: "Advar en bruker.",
-	},
-
-	DMPermission:             utils.Ref(false),
-	DefaultMemberPermissions: json.NewNullablePtr(discord.PermissionKickMembers),
-	Options: []discord.ApplicationCommandOption{
-		discord.ApplicationCommandOptionUser{
-			Name: "user",
-			NameLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "bruker",
-			},
-			Description: "The user to warn.",
-			DescriptionLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "Brukeren du vil advare.",
-			},
-			Required: true,
-		},
-
-		discord.ApplicationCommandOptionString{
-			Name: "reason",
-			NameLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "aarsak",
-			},
-			Description: "The reason for the warning.",
-			DescriptionLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "Årsaken til advarselen.",
-			},
-			Required: true,
-		},
-
-		discord.ApplicationCommandOptionFloat{
-			Name: "severity",
-			NameLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "alvorlighet",
-			},
-			Description: "The severity of the warning.",
-			DescriptionLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "Alvorlighetsgraden til advarselen.",
-			},
-			Required: false,
-			MinValue: utils.Ref(0.0),
-			MaxValue: utils.Ref(10.0),
-		},
-
-		discord.ApplicationCommandOptionBool{
-			Name: "silent",
-			NameLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "stille",
-			},
-			Description: "Whether the warning should be silent / logged without messaging the user",
-			DescriptionLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "Om advarselen skal være stille / lagres uten å varsle brukeren",
-			},
-			Required: false,
-		},
-	},
-}
-
-func WarnHandler(e *handler.CommandEvent) error {
-	utils.LogInteraction("warn", e)
-
-	ctx := context.Background()
-	data := e.SlashCommandInteractionData()
-
-	user := data.User("user")
-	reason := data.String("reason")
-	severity, severityIsSet := data.OptFloat("severity")
-	silent, silentIsSet := data.OptBool("silent")
-
-	if !severityIsSet {
-		severity = 1.0
-	}
-	if !silentIsSet {
-		silent = false
-	}
-
-	guild, ok := e.Guild()
-	if !ok {
-		slog.Warn("No guild id found in event.", "guild", guild)
-		return interactions.ErrEventNoGuildID
-	}
-
-	slog.DebugContext(
-		ctx, "Received /warn command.",
-		"user", user.Username,
-		"guild", guild.Name,
-		"moderator", e.User().Username,
-	)
-
-	inf, err := model.CreateInfraction(guild.ID, user.ID, e.User().ID, reason, severity, silent)
-	if err != nil {
-		_ = e.CreateMessage(
-			discord.NewMessageCreateBuilder().
-				SetEphemeral(true).
-				SetContent("Failed to create infraction.").Build(),
-		)
-		return fmt.Errorf("failed to create infraction: %w", err)
-	}
-
-	slog.DebugContext(ctx, "Created infraction.", "infraction", inf.Sqid())
-
-	embed := discord.NewEmbedBuilder().
-		SetTitlef(`Warning in "%s"`, guild.Name).
-		SetDescription(inf.Reason).
-		SetColor(severityToColor(inf.Weight)).
-		SetTimestamp(inf.Timestamp)
-
-	slog.DebugContext(ctx, "Created embed.")
-
-	failedToSend := false
-	if !inf.Silent {
-		channel, err := e.Client().Rest().CreateDMChannel(user.ID)
-		if err != nil {
-			failedToSend = true
-		}
-		_, err = e.Client().Rest().CreateMessage(
-			channel.ID(), discord.NewMessageCreateBuilder().
-				SetEmbeds(embed.Build()).
-				Build(),
-		)
-		if err != nil {
-			failedToSend = true
-		}
-	}
-
-	message := discord.NewMessageCreateBuilder().
-		SetContentf(
-			"## %s %s.",
-			utils.Iif(
-				inf.Silent, "Silent warning created for",
-				utils.Iif(failedToSend, "Failed to send warning to", "Warning sent to"),
-			),
-			user.Mention(),
-		).
-		SetEmbeds(embed.Build()).
-		Build()
-
-	guildSettings, err := model.GetGuildSettings(guild.ID)
-	if err == nil && guildSettings.ModeratorChannel != 0 {
-		_, err = e.Client().Rest().CreateMessage(guildSettings.ModeratorChannel, message)
-		if err != nil {
-			slog.Error(
-				"Failed to send warning to moderator channel.",
-				"err", err,
-				"guildID", guild.ID,
-				"channelID", guildSettings.ModeratorChannel,
-				"userID", user.ID,
-			)
-		}
-	}
-
-	return e.CreateMessage(
-		discord.NewMessageCreateBuilder().
-			SetEphemeral(true).
-			SetContentf(
-				"Warning created for %s.",
-				user.Mention(),
-			).Build(),
-	)
-}
-
-func severityToColor(severity float64) int {
-	if severity >= 3.0 {
-		return 0xFF0000
-	}
-	if severity >= 1.0 {
-		return 0xFF9100
-	}
-	return 0xFFFF00
-}
-
-// ░█░█░█▀▀░█▀▀░█▀▄░░░▀█▀░█▀█░█▀▀░█▀▄░█▀█░█▀▀░▀█▀░▀█▀░█▀█░█▀█░█▀▀
-// ░█░█░▀▀█░█▀▀░█▀▄░░░░█░░█░█░█▀▀░█▀▄░█▀█░█░░░░█░░░█░░█░█░█░█░▀▀█
-// ░▀▀▀░▀▀▀░▀▀▀░▀░▀░░░▀▀▀░▀░▀░▀░░░▀░▀░▀░▀░▀▀▀░░▀░░▀▀▀░▀▀▀░▀░▀░▀▀▀
-
-// UserInfractionsCommand lets users view their own infractions.
-var UserInfractionsCommand = discord.SlashCommandCreate{
-	Name: "warnings",
-	NameLocalizations: map[discord.Locale]string{
-		discord.LocaleNorwegian: "advarsler",
-	},
-	Description: "View your warnings.",
-	DescriptionLocalizations: map[discord.Locale]string{
-		discord.LocaleNorwegian: "Se advarslene dine.",
-	},
-
-	DMPermission: utils.Ref(false),
-}
-
-func UserInfractionsHandler(e *handler.CommandEvent) error {
-	utils.LogInteraction("warnings", e)
-
-	user := e.User()
-	guild, ok := e.Guild()
-	if !ok {
-		slog.Warn("No guild id found in event.", "guild", guild)
-		return interactions.ErrEventNoGuildID
-	}
-
-	message, err := getUserInfractionsAndMakeMessage(false, &guild, &user)
-	if err != nil {
-		slog.Error("Error occurred getting infractions", "err", err)
-	}
-
-	return e.CreateMessage(message.Build())
-}
-
-func UserInfractionButtonHandler(e *handler.ComponentEvent) error {
-	utils.LogInteraction("warnings button", e)
-	offsetStr := e.Vars["offset"]
-	offset, err := strconv.Atoi(offsetStr)
-	if err != nil {
-		return fmt.Errorf("failed to parse offset: %w", err)
-	}
-
-	user := e.User()
-	guild, ok := e.Guild()
-	if !ok {
-		return interactions.ErrEventNoGuildID
-	}
-
-	mcb, mub, err := getUserInfractionsAndUpdateMessage(false, offset, &guild, &user)
-	if err != nil {
-		slog.Error("Error occurred getting infractions", "err", err)
-	}
-	if mcb != nil {
-		return e.CreateMessage(mcb.Build())
-	}
-	return e.UpdateMessage(mub.Build())
-}
-
-//░▀█▀░█▀█░█▀▀░█▀▄░█▀█░█▀▀░▀█▀░▀█▀░█▀█░█▀█░█▀▀░░░▄▀░░█▄█░█▀█░█▀▄░▀▄░
-//░░█░░█░█░█▀▀░█▀▄░█▀█░█░░░░█░░░█░░█░█░█░█░▀▀█░░░█░░░█░█░█░█░█░█░░█░
-//░▀▀▀░▀░▀░▀░░░▀░▀░▀░▀░▀▀▀░░▀░░▀▀▀░▀▀▀░▀░▀░▀▀▀░░░░▀░░▀░▀░▀▀▀░▀▀░░▀░░
-
-// InfractionsCommand is a set of subcommands to manage infractions.
-var InfractionsCommand = discord.SlashCommandCreate{
-	Name: "infractions",
-	NameLocalizations: map[discord.Locale]string{
-		discord.LocaleNorwegian: "advarsler",
-	},
-	Description: "View a user's warnings.",
-	DescriptionLocalizations: map[discord.Locale]string{
-		discord.LocaleNorwegian: "Se en brukers advarsler.",
-	},
-
-	DMPermission:             utils.Ref(false),
-	DefaultMemberPermissions: json.NewNullablePtr(discord.PermissionKickMembers),
-	Options: []discord.ApplicationCommandOption{
-		discord.ApplicationCommandOptionSubCommand{
-			Name: "list",
-			NameLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "liste",
-			},
-			Description: "View a user's warnings. (NB: Response visible to all)",
-			DescriptionLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "Se en brukers advarsler.",
-			},
-			Options: []discord.ApplicationCommandOption{
-				discord.ApplicationCommandOptionUser{
-					Name: "user",
-					NameLocalizations: map[discord.Locale]string{
-						discord.LocaleNorwegian: "bruker",
-					},
-					Description: "The user to view warnings for.",
-					DescriptionLocalizations: map[discord.Locale]string{
-						discord.LocaleNorwegian: "Brukeren du vil se advarsler for.",
-					},
-					Required: false,
-				},
-				discord.ApplicationCommandOptionString{
-					Name: "user-id",
-					NameLocalizations: map[discord.Locale]string{
-						discord.LocaleNorwegian: "bruker-id",
-					},
-					Description: "The ID of the user user to view warnings for.",
-					DescriptionLocalizations: map[discord.Locale]string{
-						discord.LocaleNorwegian: "ID-en til brukeren du vil se advarsler for.",
-					},
-					Required: false,
-				},
-			},
-		},
-
-		discord.ApplicationCommandOptionSubCommand{
-			Name: "remove",
-			NameLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "fjern",
-			},
-			Description: "Remove a user's warning.",
-			DescriptionLocalizations: map[discord.Locale]string{
-				discord.LocaleNorwegian: "Fjern en brukers advarsel.",
-			},
-			Options: []discord.ApplicationCommandOption{
-				discord.ApplicationCommandOptionString{
-					Name: "infraction-id",
-					NameLocalizations: map[discord.Locale]string{
-						discord.LocaleNorwegian: "advarsels-id",
-					},
-					Description: "The id of the infraction to remove.",
-					DescriptionLocalizations: map[discord.Locale]string{
-						discord.LocaleNorwegian: "ID-en til advarselen du vil fjerne.",
-					},
-					Required: true,
-				},
-			},
-		},
-	},
-}
-
-// InfractionsListHandler handles the `/infractions list` command.
-func InfractionsListHandler(e *handler.CommandEvent) error {
-	utils.LogInteraction("infractions list", e)
-
-	slog.Info("interaction `/infractions list` called.")
-	data := e.SlashCommandInteractionData()
-	user, hasUser := data.OptUser("user")
-	userIDString, hasUserID := data.OptString("user-id")
-
-	if !hasUser && !hasUserID {
-		return e.CreateMessage(
-			discord.NewMessageCreateBuilder().
-				SetContent("You must specify either a user or a user ID.").
-				SetEphemeral(true).
-				Build(),
-		)
-	}
-
-	if hasUser && hasUserID {
-		return e.CreateMessage(
-			discord.NewMessageCreateBuilder().
-				SetContent("You can only specify either a user or a user ID.").
-				SetEphemeral(true).
-				Build(),
-		)
-	}
-
-	if !hasUser {
-		userID, err := snowflake.Parse(userIDString)
-		if err != nil {
-			_ = e.CreateMessage(
-				discord.NewMessageCreateBuilder().
-					SetContent("Failed to parse user id.").
-					SetEphemeral(true).
-					Build(),
-			)
-			return fmt.Errorf("failed to parse user id: %w", err)
-		}
-
-		userRef, err := e.Client().Rest().GetUser(userID)
-		if err != nil || userRef == nil {
-			user = discord.User{
-				ID:       userID,
-				Username: "unknown_user",
-			}
-		} else {
-			user = *userRef
-		}
-	}
-	guild, ok := e.Guild()
-	if !ok {
-		slog.Warn("No guild id found in event.", "guild", guild)
-		return interactions.ErrEventNoGuildID
-	}
-
-	message, err := getUserInfractionsAndMakeMessage(true, &guild, &user)
-	if err != nil {
-		slog.Error("Error occurred getting infractions", "err", err)
-	}
-
-	return e.CreateMessage(message.Build())
-}
-
-func InfractionsRemoveHandler(e *handler.CommandEvent) error {
-	utils.LogInteraction("infractions remove", e)
-
-	data := e.SlashCommandInteractionData()
-	infID := data.String("infraction-id")
-	guild, ok := e.Guild()
-	if !ok {
-		slog.Warn("No guild id found in event.", "guild", guild)
-		return interactions.ErrEventNoGuildID
-	}
-
-	err := model.DeleteInfractionBySqid(infID)
-	if err != nil {
-		return e.CreateMessage(
-			discord.NewMessageCreateBuilder().
-				SetEphemeral(true).
-				SetContent("Failed to delete infraction.").
-				Build(),
-		)
-	}
-
-	return e.CreateMessage(
-		discord.NewMessageCreateBuilder().
-			SetEphemeral(true).
-			SetContent("Infraction deleted.").
-			Build(),
-	)
-}
-
-func InfractionsListComponentHandler(e *handler.ComponentEvent) error {
-	utils.LogInteraction("infractions list component", e)
-
-	parentIx := e.Message.Interaction
-	if parentIx == nil {
-		return fmt.Errorf("no parent interaction found")
-	}
-
-	guild, isGuild := e.Guild()
-	if !isGuild {
-		return interactions.ErrEventNoGuildID
-	}
-	offset, err := strconv.Atoi(e.Vars["offset"])
-	if err != nil {
-		return fmt.Errorf("failed to parse offset: %w", err)
-	}
-	userID, err := snowflake.Parse(e.Vars["userID"])
-	if err != nil {
-		return fmt.Errorf("failed to parse user id: %w", err)
-	}
-
-	user, err := e.Client().Rest().GetUser(userID)
-	if err != nil {
-		return fmt.Errorf("failed to get user: %w", err)
-	}
-
-	if e.User().ID != parentIx.User.ID {
-		return e.CreateMessage(
-			discord.NewMessageCreateBuilder().
-				SetAllowedMentions(&discord.AllowedMentions{}).
-				SetContent("You can only paginate responses from your own commands.").
-				SetEphemeral(true).
-				Build(),
-		)
-	}
-
-	mcb, mub, err := getUserInfractionsAndUpdateMessage(false, offset, &guild, user)
-	if err != nil {
-		slog.Error("Error occurred getting infractions", "err", err)
-	}
-	if mcb != nil {
-		return e.CreateMessage(mcb.Build())
-	}
-	return e.UpdateMessage(mub.Build())
-}
-
-//░█░█░▀█▀░▀█▀░█░░░▀█▀░▀█▀░█░█░░░█▀▀░▀█▀░█▀▄░█░█░█▀▀░▀█▀░█▀▀░░░█░█▀▀░█░█░█▀█░█▀▀░█▀▀
-//░█░█░░█░░░█░░█░░░░█░░░█░░░█░░░░▀▀█░░█░░█▀▄░█░█░█░░░░█░░▀▀█░▄▀░░█▀▀░█░█░█░█░█░░░▀▀█
-//░▀▀▀░░▀░░▀▀▀░▀▀▀░▀▀▀░░▀░░░▀░░░░▀▀▀░░▀░░▀░▀░▀▀▀░▀▀▀░░▀░░▀▀▀░▀░░░▀░░░▀▀▀░▀░▀░▀▀▀░▀▀▀
 
 type userInfractions struct {
 	Embeds        []discord.Embed
@@ -620,6 +152,16 @@ func createInfractionEmbeds(infractions []model.Infraction, guildSettings *model
 		embeds = append(embeds, embed.Build())
 	}
 	return embeds
+}
+
+func severityToColor(severity float64) int {
+	if severity >= 3.0 {
+		return 0xFF0000
+	}
+	if severity >= 1.0 {
+		return 0xFF9100
+	}
+	return 0xFFFF00
 }
 
 func severityToDots(severity float64) string {
